ohif-301: add sorting to work list series
