+++ extensions/measurement-tracking/src/panels/PanelStudyBrowserTracking/PanelStudyBrowserTracking.jsx
@@ -227,10 +227,10 @@ function PanelStudyBrowserTracking({
     );
     const updatedExpandedStudyInstanceUIDs = shouldCollapseStudy
       ? [
-          ...expandedStudyInstanceUIDs.filter(
-            stdyUid => stdyUid !== StudyInstanceUID
-          ),
-        ]
+        ...expandedStudyInstanceUIDs.filter(
+          stdyUid => stdyUid !== StudyInstanceUID
+        ),
+      ]
       : [...expandedStudyInstanceUIDs, StudyInstanceUID];
 
     setExpandedStudyInstanceUIDs(updatedExpandedStudyInstanceUIDs);
@@ -304,7 +304,7 @@ function PanelStudyBrowserTracking({
           SeriesInstanceUID: displaySet.SeriesInstanceUID,
         });
       }}
-      onClickThumbnail={() => {}}
+      onClickThumbnail={() => { }}
       onDoubleClickThumbnail={onDoubleClickThumbnailHandler}
       activeDisplaySetInstanceUID={activeDisplaySetInstanceUID}
     />
@@ -370,11 +370,11 @@ function _mapDisplaySets(
     const viewportIdentificator = isSingleViewport
       ? []
       : viewports.reduce((acc, viewportData, index) => {
-          if (viewportData.displaySetInstanceUID === ds.displaySetInstanceUID) {
-            acc.push(_viewportLabels[index]);
-          }
-          return acc;
-        }, []);
+        if (viewportData.displaySetInstanceUID === ds.displaySetInstanceUID) {
+          acc.push(_viewportLabels[index]);
+        }
+        return acc;
+      }, []);
 
     const array =
       componentType === 'thumbnailTracked'
@@ -444,20 +444,9 @@ function _createStudyBrowserTabs(
   const allStudies = [];
 
   studyDisplayList.forEach(study => {
-    const displaySetsForStudy = displaySets.filter(
+    const displaySetsForStudy = utils.sortBySeriesDate(displaySets.filter(
       ds => ds.StudyInstanceUID === study.studyInstanceUid
-    );
-
-    displaySetsForStudy.sort((a,b) => {
-      if (a.seriesNumber !== b.seriesNumber) {
-        return a.seriesNumber - b.seriesNumber;
-      }
-
-      const seriesDateA = Date.parse(a.seriesDate);
-      const seriesDateB = Date.parse(b.seriesDate);
-
-      return seriesDateA - seriesDateB;
-    });
+    ));
 
     const tabStudy = Object.assign({}, study, {
       displaySets: displaySetsForStudy,
